Restructured with identical functionality, isolate DB, schema, or application dependent code in logtools/ext_db.py
